+++ accounts/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class RegistrationLoginConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "registration_login"
+    name = "accounts"
