+++ accounts/forms.py
@@ -3,10 +3,8 @@ from django.contrib.auth.forms import UserCreationForm
 from django import forms
 from django.contrib.auth.models import User
 
+
 class CreateUserForm(UserCreationForm):
     class Meta:
         model = User
-        fields = ['username', 'email','password1','password2']
-
-
-
+        fields = ["username", "email", "password1", "password2"]
