+++ accounts/templates/registration_login/login.html
@@ -28,11 +28,9 @@
                   </div>
                   <input type="password" name="password" placeholder="Hasło..." class="form-control" >
                 </div>
-
                 {% for message in messages%}
                     <p id='messages'>{{message}}</p>
                 {% endfor %}
-
                 <div class="form-group">
                   <button type="submit" class="btn btn-primary btn-block"> Zaloguj </button>
                 </div>
