--- registration_login/urls.py
@@ -1,10 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-
-    path('register/', views.registerPage, name="register"),
-    path('login/', views.loginPage, name="login"),
-    path('logout/', views.logoutUser, name="logout"),
-
-]
