+++ weather_portal/settings.py
@@ -40,7 +40,7 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "registration_login"
+    "accounts",
 ]
 
 MIDDLEWARE = [
